requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.8.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
dataclasses-json==0.5.7
Deprecated==1.2.14
distlib==0.3.8
faiss-cpu==1.7.4
filelock==3.14.0
fitparse==1.2.0
Flask==2.3.2
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.130
langsmith==0.0.46
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.7
numpy==1.23.5
openai==0.27.6
packaging==23.1
pandas==2.0.2
platformdirs==4.2.1
pycparser==2.21
pydantic==1.10.7
PyGithub==1.58.2
PyJWT==2.7.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
SQLAlchemy==1.4.48
tenacity==8.2.2
tqdm==4.65.0
twilio==8.2.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.26.1
Werkzeug==2.3.3
wrapt==1.15.0
yarl==1.9.2
